Move implementation inside of PersistentList
