+++ subprojects/data-structures/src/main/java/org/gradle/internal/collect/PersistentList.java
@@ -19,19 +19,82 @@ package org.gradle.internal.collect;
 import org.gradle.internal.Cast;
 
 import javax.annotation.CheckReturnValue;
+import java.util.Objects;
 import java.util.function.Consumer;
 
-public interface PersistentList<T> {
-    static <T> PersistentList<T> of() {
-        return Cast.uncheckedCast(Nil.NIL);
+public abstract class PersistentList<T> {
+    public static <T> PersistentList<T> of() {
+        return Cast.uncheckedNonnullCast(NIL);
     }
 
-    static <T> PersistentList<T> of(T first) {
+    public static <T> PersistentList<T> of(T first) {
         return PersistentList.<T>of().plus(first);
     }
 
-    void forEach(Consumer<? super T> consumer);
+
+    public abstract void forEach(Consumer<? super T> consumer);
 
     @CheckReturnValue
-    PersistentList<T> plus(T element);
+    public abstract PersistentList<T> plus(T element);
+
+    private PersistentList() {}
+
+    private static final PersistentList<Object> NIL = new PersistentList<Object>() {
+        @Override
+        public void forEach(Consumer<? super Object> consumer) {
+        }
+
+        @Override
+        public PersistentList<Object> plus(Object element) {
+            return new Cons<>(element, this);
+        }
+
+        @Override
+        public String toString() {
+            return "Nil";
+        }
+    };
+
+    private static class Cons<T> extends PersistentList<T> {
+        private final T head;
+        private final PersistentList<T> tail;
+
+        public Cons(T head, PersistentList<T> tail) {
+            this.head = head;
+            this.tail = tail;
+        }
+
+        @Override
+        public void forEach(Consumer<? super T> consumer) {
+            consumer.accept(head);
+            tail.forEach(consumer);
+        }
+
+        @Override
+        public PersistentList<T> plus(T element) {
+            return new Cons<>(element, this);
+        }
+
+        @Override
+        public boolean equals(Object o) {
+            if (this == o) {
+                return true;
+            }
+            if (o == null || getClass() != o.getClass()) {
+                return false;
+            }
+            Cons<?> cons = (Cons<?>) o;
+            return head.equals(cons.head) && tail.equals(cons.tail);
+        }
+
+        @Override
+        public int hashCode() {
+            return Objects.hash(head, tail);
+        }
+
+        @Override
+        public String toString() {
+            return head + " : " + tail;
+        }
+    }
 }
